refactor: split the row codec and the legacy-layout store tests out
Both files crossed the 500-line cap on this branch. `store-codec.ts` splits at
the scope seam its own header names: manifest-level concerns (lock retry,
corrupt/unsupported recovery, read-merge-write) stay, and the per-row
"one JSON entry to a Task" half moves to `store-codec-rows.ts`. The store's
legacy-`.kobe` load tests move to `store-legacy-layout.test.ts`, where the
input under test is the layout rather than the bytes.
